Split runtime capability tests for backend size gate
