+++ pkg/workers/executor/agent_test.go
@@ -84,12 +84,6 @@ func withAgentWorkingDirectory(workingDirectory string) func(*interfaces.Worksta
 	}
 }
 
-func withAgentWorkingDirectoryAuthored(authored bool) func(*interfaces.WorkstationExecutionRequest) {
-	return func(req *interfaces.WorkstationExecutionRequest) {
-		req.WorkingDirectoryAuthored = authored
-	}
-}
-
 func withAgentModelOperation(operation string, bindings []interfaces.ResolvedModelOperationBinding) func(*interfaces.WorkstationExecutionRequest) {
 	return func(req *interfaces.WorkstationExecutionRequest) {
 		req.ModelOperation = operation
@@ -169,63 +163,6 @@ func TestInferenceRequestForExecutionRequest_ForwardsModelOperationContract(t *t
 	got.ModelBindings[0].Content[0].Text = "changed"
 	if req.ModelBindings[0].Content[0].Text != "hello" {
 		t.Fatalf("request bindings mutated original execution request: %#v", req.ModelBindings)
-	}
-}
-
-func TestInferenceRequestForExecutionRequest_DefaultWorkingDirectoryDoesNotRequireRunnerCapability(t *testing.T) {
-	req := testAgentRequest(
-		interfaces.WorkDispatch{
-			DispatchID:      "d-default-workdir",
-			TransitionID:    "t-default-workdir",
-			WorkerType:      "worker-a",
-			WorkstationName: "review",
-		},
-		withAgentPrompts("System prompt", "Review"),
-		withAgentWorkingDirectory("/tmp/runtime-session"),
-	)
-
-	got := inferenceRequestForExecutionRequest(req, &interfaces.WorkerConfig{
-		Model:         "gemini-1.5-pro",
-		ModelProvider: interfaces.RunnerIDGemini,
-	}, nil)
-
-	for _, capability := range got.RequiredOptionalCapabilities {
-		if capability == interfaces.RunnerOptionalCapabilityWorkingDirectory {
-			t.Fatalf("capabilities = %#v, want default working directory omitted", got.RequiredOptionalCapabilities)
-		}
-	}
-	if got.WorkingDirectory != "/tmp/runtime-session" {
-		t.Fatalf("working directory = %q, want forwarded runtime path", got.WorkingDirectory)
-	}
-}
-
-func TestInferenceRequestForExecutionRequest_AuthoredWorkingDirectoryRequiresRunnerCapability(t *testing.T) {
-	req := testAgentRequest(
-		interfaces.WorkDispatch{
-			DispatchID:      "d-authored-workdir",
-			TransitionID:    "t-authored-workdir",
-			WorkerType:      "worker-a",
-			WorkstationName: "review",
-		},
-		withAgentPrompts("System prompt", "Review"),
-		withAgentWorkingDirectory("/tmp/authored"),
-		withAgentWorkingDirectoryAuthored(true),
-	)
-
-	got := inferenceRequestForExecutionRequest(req, &interfaces.WorkerConfig{
-		Model:         "gemini-1.5-pro",
-		ModelProvider: interfaces.RunnerIDGemini,
-	}, nil)
-
-	found := false
-	for _, capability := range got.RequiredOptionalCapabilities {
-		if capability == interfaces.RunnerOptionalCapabilityWorkingDirectory {
-			found = true
-			break
-		}
-	}
-	if !found {
-		t.Fatalf("capabilities = %#v, want authored working directory capability", got.RequiredOptionalCapabilities)
 	}
 }
